+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This is a changelog for **ZachranObed** application.
 
 ### Changed
 - **ZOB-105** Add non-prod firebase projects.
+- **ZOB-92** Update toolbar in Profile screen.
 
 ### Fixed